fix(#2650): check_row_count_spike can be called at its documented default
`check_row_count_spike(..., *, exclude_run_id: int | None = None)` could not
be called at the default its own signature declares. The comparison query
filtered the current run out of its history with

    AND (%(exclude_id)s IS NULL OR run_id != %(exclude_id)s)

psycopg3 dedups the repeated named parameter into one `$n`. At `None` it is
sent untyped (OID 0) and its only type-determining context is a `NullTest`,
which constrains nothing — `run_id != $n` does not rescue it, because operator
resolution runs after parameter-type resolution. Measured against the real
query on dev Postgres 17:

    exclude_run_id=None -> AmbiguousParameter: could not determine data type
                           of parameter $2
    exclude_run_id=1    -> OK

Latent, not live: both production callers (app/workers/scheduler.py:2397,
:2444) pass `tracker.run_id`, declared `int` and never None.

Cast `::bigint` at both occurrences — `job_runs.run_id` confirmed `bigint`
from information_schema, not assumed. A partial cast is NOT enough and that is
now measured: casting only the `!=` occurrence still raises AmbiguousParameter.

Adds a db-tier twin. The existing MagicMock file drives a mocked cursor that
records parameters and sends nothing, so it cannot observe a plan-time type
failure while asserting the exact failing path five times over. The new tests
seed two prior runs and assert the exclusion FLIPS the flagged verdict, because
`COALESCE(%(exclude_id)s, -1)` would also plan while destroying the
self-comparison guard the clause exists for.

Sweep: this was the last live occurrence of the shape in app/ + scripts/ (grep
recorded in the prevention-log entry); the only other hits are prose.

⚠ Skill gap left open and recorded in the log: the sql-correctness bullet is
still written as "nullable-FILTER param", the narrow framing #2647 identified
as the reason this class escapes. Edit on .claude/skills/** was refused in the
autonomy worktree, so the broadening landed in the prevention log only.

Closes #2650. Refs #1961. Refs #2647.

+++ app/services/sync_orchestrator/row_count_spikes.py
@@ -61,7 +61,18 @@ def check_row_count_spike(
             WHERE job_name = %(name)s
               AND status = 'success'
               AND row_count IS NOT NULL
-              AND (%(exclude_id)s IS NULL OR run_id != %(exclude_id)s)
+              -- ::bigint at EVERY occurrence, matching job_runs.run_id.
+              -- psycopg3 dedups the repeated named param into one $n, and
+              -- `$n IS NULL` is a NullTest that determines no type, so the
+              -- default exclude_run_id=None binds as an untyped NULL and the
+              -- whole statement fails to plan with AmbiguousParameter.
+              -- `run_id != $n` does NOT rescue it. See prevention-log
+              -- §"A nullable SQL param whose only type-determining use is a
+              -- NULL test must be cast" (#1961/#2647/#2650).
+              AND (
+                  %(exclude_id)s::bigint IS NULL
+                  OR run_id != %(exclude_id)s::bigint
+              )
             ORDER BY started_at DESC
             LIMIT 1
             """,
